booking_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
aiokafka==0.11.0
redis==5.0.1
python-dotenv==1.0.0
